+++ Assets/Scripts/Navigation/StartPageFunctions.cs
